feat: implement multi contract batch call
