+++ src/utils/dao/getContractNameByCodeId.ts
@@ -1,0 +1,6 @@
+import { customQueries } from '@ixo/impactxclient-sdk'
+import { chainNetwork } from 'hooks/configs'
+
+export const getContractNameByCodeId = (codeId: number): string => {
+  return customQueries.contract.getContractCodes(chainNetwork).find(({ code }) => code === codeId)?.name ?? ''
+}
